+++ src/maicos/core/planar.py
@@ -227,7 +227,7 @@ class ProfilePlanarBase(PlanarBase, ProfileBase):
         if self.sym and self.refgroup is None:
             raise ValueError("For symmetrization the `refgroup` argument is required.")
 
-        logging.info(f"Computing {self.grouping} profile along {'XYZ'[self.dim]}-axes.")
+        logging.info(f"""Profile along {"xyz"[self.dim]}-axis normal to the plane.""")
 
     def _compute_histogram(
         self, positions: np.ndarray, weights: np.ndarray | None = None